fix inline i386 assembler to use the correct register sizes
change the windows implementation of QMutex to use q_atomic_*

done with Brad

[git-p4: depot-paths = "//depot/qt/main/": change = 130963]

+++ src/core/thread/qmutex_p.h
@@ -45,7 +45,7 @@ class QMutexPrivate {
 public:
     bool recursive;
 
-    unsigned long owner;
+    int owner;
     unsigned int count;
     QAtomic waiters;
